% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Bin a vector}
\usage{
bin(time, values, bin.vec, bin.fun = mean)
}
\arguments{
\item{time}{time vector}

\item{values}{value vector}

\item{bin.vec}{edges of the bins}

\item{bin.fun}{function to bin with}
}
\value{
data frame
}
\description{
Bin a vector
}
