% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_rabd660670}
\alias{calibrate_rabd660670}
\title{Convert RABD to chl a concentration}
\usage{
calibrate_rabd660670(rabd660670, geoClass)
}
\arguments{
\item{rabd660670}{RABD 660-670}

\item{geoClass}{Geomorphic class of the lake}
}
\value{
chla concentrations in ug/g
}
\description{
Convert RABD to chl a concentration
}
