% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{chla_flux_to_tibble}
\alias{chla_flux_to_tibble}
\title{Create a tibble from the flux output}
\usage{
chla_flux_to_tibble(tib)
}
\arguments{
\item{tib}{the output of estimate_chla_flux_lipd or estimate_chla_flux}
}
\value{
a tibble of output data
}
\description{
Create a tibble from the flux output
}
