% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{estimate_chla_flux}
\alias{estimate_chla_flux}
\title{Do the chl a flux calculations}
\usage{
estimate_chla_flux(depth, time, rabd660670, geoClass, smooth = TRUE)
}
\arguments{
\item{depth}{depth in cm}

\item{time}{time in BP or AD}

\item{rabd660670}{RABD 660 670 vector}

\item{geoClass}{geoclass}

\item{smooth}{optionally smooth out sedrate curve to avoid unrealistic sharp jumps. NA (default) will do nothing. Specify degrees of freedom in spline to smooth. TRUE will use df = span in cm divided by 2}
}
\value{
a tibble with output data
}
\description{
Do the chl a flux calculations
}
