% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{estimate_chla_flux_lipd}
\alias{estimate_chla_flux_lipd}
\title{Do the chl a flux calculations from a lipd object}
\usage{
estimate_chla_flux_lipd(ts, smooth = TRUE)
}
\arguments{
\item{ts}{a lipd-ts-tibble object, or somethign that can be converted to one.}

\item{smooth}{optionally smooth out sedrate curve to avoid unrealistic sharp jumps. NA (default) will do nothing. Specify degrees of freedom in spline to smooth. TRUE will use df = span in cm divided by 2}
}
\value{
a tibble with output data
}
\description{
Do the chl a flux calculations from a lipd object
}
