% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{estimate_dbd}
\alias{estimate_dbd}
\title{Estimate wet bulk density from depth}
\usage{
estimate_dbd(dblf)
}
\arguments{
\item{dblf}{depth below lake floor in cm}
}
\value{
dry bulk density in g/cm3
}
\description{
Estimate wet bulk density from depth
}
\examples{
dbd <- estimate_dbd(seq(1,150))
}
