% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedrate.R
\name{estimate_sed_rate_from_median}
\alias{estimate_sed_rate_from_median}
\title{Simple sed rate estimation from a median time estimate and depth}
\usage{
estimate_sed_rate_from_median(time, depth, smooth = NA)
}
\arguments{
\item{time}{vector of age or year values}

\item{depth}{vector of depth values}

\item{smooth}{optionally smooth out sedrate curve to avoid unrealistic sharp jumps. NA (default) will do nothing. Specify degrees of freedom in spline to smooth. TRUE will use df = span in cm divided by 2}
}
\value{
estimate sed rate
}
\description{
Simple sed rate estimation from a median time estimate and depth
}
