% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{estimate_wbd}
\alias{estimate_wbd}
\title{Estimate wet bulk density from depth}
\usage{
estimate_wbd(dblf)
}
\arguments{
\item{dblf}{depth below lake floor in cm}
}
\value{
wet bulk density in g/cm3
}
\description{
Estimate wet bulk density from depth
}
\examples{
wbd <- estimate_wbd(seq(1,150))
}
