% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_flux}
\alias{plot_flux}
\title{Plot the key datasets in the flux calculation}
\usage{
plot_flux(tib)
}
\arguments{
\item{tib}{the output of estimate_chla_flux_lipd() or estimate_chla_flux()}
}
\value{
a ggplot object
}
\description{
Plot the key datasets in the flux calculation
}
